+++ src/test/java/io/github/dlbbld/ashlarchess/report/TestReportPrintoutDerivesFromObjectModel.java
@@ -200,6 +200,6 @@ class TestReportPrintoutDerivesFromObjectModel {
 
   private static Board loadCorpusBoard(String pgnName) {
     final PgnTest pgnTest = PgnTestCaseCatalog.findPgnTestPgnNotListed(pgnName);
-    return PgnUtility.calculateBoard(pgnTest.getFolderPath(), pgnName);
+    return PgnUtility.toBoard(pgnTest.getFolderPath(), pgnName);
   }
 }