+++ src/test/java/io/github/dlbbld/ashlarchess/test/tools/FenFromPgn.java
@@ -45,7 +45,7 @@ public final class FenFromPgn {
     final String fileName = Nulls.toString(Nulls.getFileName(pgnPath));
 
     final PgnGame pgnGame = LenientPgnParser.parsePath(folder, fileName);
-    final Board board = PgnUtility.calculateBoard(pgnGame);
+    final Board board = PgnUtility.toBoard(pgnGame);
     System.out.println(board.getFen());
   }
 }